OCSADV-251: Moving web-only stuff to new web module.
