+++ app/itc/build.sbt
@@ -51,6 +51,7 @@ def common(version: Version) = AppConfig(
     BundleSpec("edu.gemini.util.log.extras",             version),
     BundleSpec("edu.gemini.itc",                         version),
     BundleSpec("edu.gemini.itc.shared",                  version),
+    BundleSpec("edu.gemini.itc.web",                     version),
     BundleSpec("org.apache.commons.io",                  Version(2, 0, 1)),
     BundleSpec("org.ops4j.pax.web.pax-web-extender-war", Version(1, 1, 13)),
     BundleSpec("org.ops4j.pax.web.pax-web-jetty-bundle", Version(1, 1, 13)),
