+++ bundle/edu.gemini.itc.web/src/main/java/edu/gemini/itc/web/servlets/ITCMichelleServlet.java
@@ -1,20 +1,20 @@
-package edu.gemini.itc.niri;
+package edu.gemini.itc.web.servlets;
 
+import edu.gemini.itc.michelle.MichelleRecipe;
 import edu.gemini.itc.shared.ITCMultiPartParser;
-import edu.gemini.itc.shared.ITCServlet;
 import edu.gemini.itc.shared.Recipe;
 
 import java.io.PrintWriter;
 
 /**
- * ITC NIRI servlet.
+ * ITC Michelle servlet.
  */
-public final class ITCniriServlet extends ITCServlet {
+public final class ITCMichelleServlet extends ITCServlet {
     public static final String VERSION = "4.2";
     public static final String TITLE = "Gemini Integration Time Calculator";
-    public static final String INSTRUMENT = "NIRI";
+    public static final String INSTRUMENT = "Michelle";
 
-    public ITCniriServlet() {
+    public ITCMichelleServlet() {
         super();
     }
 
@@ -44,7 +44,7 @@ public final class ITCniriServlet extends ITCServlet {
      * Used by Java Web Server Administration Tool.
      */
     public String getServletInfo() {
-        return getTitle() + " " + getVersion() + " - ITCNiriServlet accepts form data and performs ITC calculation for Niri.";
+        return getTitle() + " " + getVersion() + " - ITCMichelleServlet accepts form data and performs ITC calculation for Michelle.";
     }
 
     /**
@@ -52,7 +52,7 @@ public final class ITCniriServlet extends ITCServlet {
      */
     public void writeOutput(ITCMultiPartParser mpp, PrintWriter out) {
         out.println("<a href = \"http://www.gemini.edu/sciops/instruments/integration-time-calculators/itc-help\"> Click here for help with the results page.</a>");
-        Recipe recipe = new NiriRecipe(mpp, out);
+        Recipe recipe = new MichelleRecipe(mpp, out);
         recipe.writeOutput();
     }
 }
