+++ bundle/edu.gemini.itc.web/src/main/java/edu/gemini/itc/web/servlets/ITCServlet.java
@@ -1,4 +1,7 @@
-package edu.gemini.itc.shared;
+package edu.gemini.itc.web.servlets;
+
+import edu.gemini.itc.shared.ITCMultiPartParser;
+import edu.gemini.itc.shared.ServerInfo;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
