+++ bundle/edu.gemini.itc.web/src/main/java/edu/gemini/itc/web/servlets/ITCgmosServlet.java
@@ -1,7 +1,7 @@
-package edu.gemini.itc.gmos;
+package edu.gemini.itc.web.servlets;
 
+import edu.gemini.itc.gmos.GmosRecipe;
 import edu.gemini.itc.shared.ITCMultiPartParser;
-import edu.gemini.itc.shared.ITCServlet;
 import edu.gemini.itc.shared.Recipe;
 
 import java.io.PrintWriter;
