+++ bundle/edu.gemini.itc.web/src/main/java/edu/gemini/itc/web/servlets/ITCgnirsServlet.java
@@ -1,21 +1,20 @@
-package edu.gemini.itc.acqcam;
+package edu.gemini.itc.web.servlets;
 
+import edu.gemini.itc.gnirs.GnirsRecipe;
 import edu.gemini.itc.shared.ITCMultiPartParser;
-import edu.gemini.itc.shared.ITCServlet;
 import edu.gemini.itc.shared.Recipe;
 
 import java.io.PrintWriter;
 
 /**
- * ITC Acq Cam servlet.
+ * ITC GNIRS servlet.
  */
-public final class ITCacqCamServlet
-        extends ITCServlet {
+public final class ITCgnirsServlet extends ITCServlet {
     public static final String VERSION = "4.0";
     public static final String TITLE = "Gemini Integration Time Calculator";
-    public static final String INSTRUMENT = "Acquisition Camera";
+    public static final String INSTRUMENT = "GNIRS";
 
-    public ITCacqCamServlet() {
+    public ITCgnirsServlet() {
         super();
     }
 
@@ -34,7 +33,7 @@ public final class ITCacqCamServlet
     }
 
     /**
-     * Returns the instrument name
+     * Returns the Instrument name
      */
     public String getInst() {
         return INSTRUMENT;
@@ -45,7 +44,7 @@ public final class ITCacqCamServlet
      * Used by Java Web Server Administration Tool.
      */
     public String getServletInfo() {
-        return getTitle() + " " + getVersion() + " - ITCacqCamServlet accepts form data and performs ITC calculation for the Acquisition Camera.";
+        return getTitle() + " " + getVersion() + " - ITCgnirsServlet accepts form data and performs ITC calculation for GNIRS.";
     }
 
     /**
@@ -53,7 +52,7 @@ public final class ITCacqCamServlet
      */
     public void writeOutput(ITCMultiPartParser mpp, PrintWriter out) {
         out.println("<a href = \"http://www.gemini.edu/sciops/instruments/integration-time-calculators/itc-help\"> Click here for help with the results page.</a>");
-        Recipe recipe = new AcqCamRecipe(mpp, out);
+        Recipe recipe = new GnirsRecipe(mpp, out);
         recipe.writeOutput();
     }
 }
