+++ bundle/edu.gemini.itc.web/src/main/java/edu/gemini/itc/web/servlets/ITCnifsServlet.java
@@ -1,20 +1,20 @@
-package edu.gemini.itc.gsaoi;
+package edu.gemini.itc.web.servlets;
 
+import edu.gemini.itc.nifs.NifsRecipe;
 import edu.gemini.itc.shared.ITCMultiPartParser;
-import edu.gemini.itc.shared.ITCServlet;
 import edu.gemini.itc.shared.Recipe;
 
 import java.io.PrintWriter;
 
 /**
- * ITC GSAOI servlet.
+ * ITC NIFS servlet.
  */
-public final class ITCgsaoiServlet extends ITCServlet {
+public final class ITCnifsServlet extends ITCServlet {
     public static final String VERSION = "4.2";
     public static final String TITLE = "Gemini Integration Time Calculator";
-    public static final String INSTRUMENT = "GSAOI";
+    public static final String INSTRUMENT = "NIFS";
 
-    public ITCgsaoiServlet() {
+    public ITCnifsServlet() {
         super();
     }
 
@@ -44,7 +44,7 @@ public final class ITCgsaoiServlet extends ITCServlet {
      * Used by Java Web Server Administration Tool.
      */
     public String getServletInfo() {
-        return getTitle() + " " + getVersion() + " - ITCGsaoiServlet accepts form data and performs ITC calculation for Gsaoi.";
+        return getTitle() + " " + getVersion() + " - ITCnifsServlet accepts form data and performs ITC calculation for NIFS.";
     }
 
     /**
@@ -52,7 +52,7 @@ public final class ITCgsaoiServlet extends ITCServlet {
      */
     public void writeOutput(ITCMultiPartParser mpp, PrintWriter out) {
         out.println("<a href = \"http://www.gemini.edu/sciops/instruments/integration-time-calculators/itc-help\"> Click here for help with the results page.</a>");
-        Recipe recipe = new GsaoiRecipe(mpp, out); // parses form data
+        Recipe recipe = new NifsRecipe(mpp, out);
         recipe.writeOutput();
     }
 }
