+++ bundle/edu.gemini.itc.web/src/main/java/edu/gemini/itc/web/servlets/ITCniriServlet.java
@@ -1,20 +1,20 @@
-package edu.gemini.itc.michelle;
+package edu.gemini.itc.web.servlets;
 
+import edu.gemini.itc.niri.NiriRecipe;
 import edu.gemini.itc.shared.ITCMultiPartParser;
-import edu.gemini.itc.shared.ITCServlet;
 import edu.gemini.itc.shared.Recipe;
 
 import java.io.PrintWriter;
 
 /**
- * ITC Michelle servlet.
+ * ITC NIRI servlet.
  */
-public final class ITCMichelleServlet extends ITCServlet {
+public final class ITCniriServlet extends ITCServlet {
     public static final String VERSION = "4.2";
     public static final String TITLE = "Gemini Integration Time Calculator";
-    public static final String INSTRUMENT = "Michelle";
+    public static final String INSTRUMENT = "NIRI";
 
-    public ITCMichelleServlet() {
+    public ITCniriServlet() {
         super();
     }
 
@@ -44,7 +44,7 @@ public final class ITCMichelleServlet extends ITCServlet {
      * Used by Java Web Server Administration Tool.
      */
     public String getServletInfo() {
-        return getTitle() + " " + getVersion() + " - ITCMichelleServlet accepts form data and performs ITC calculation for Michelle.";
+        return getTitle() + " " + getVersion() + " - ITCNiriServlet accepts form data and performs ITC calculation for Niri.";
     }
 
     /**
@@ -52,7 +52,7 @@ public final class ITCMichelleServlet extends ITCServlet {
      */
     public void writeOutput(ITCMultiPartParser mpp, PrintWriter out) {
         out.println("<a href = \"http://www.gemini.edu/sciops/instruments/integration-time-calculators/itc-help\"> Click here for help with the results page.</a>");
-        Recipe recipe = new MichelleRecipe(mpp, out);
+        Recipe recipe = new NiriRecipe(mpp, out);
         recipe.writeOutput();
     }
 }
